Add `bundler` parameter to STDIN adapters
A recent change in rubygems/rubygems@076aba8 is causing the bundler to attempt to reload the script, if the used Bundler version doesn't match that, specified in the Gemfile.lock. This causes issues (#3).

A quick workaround for that is to run the remote script using `bundle exec`. In order to do this, I introduced a `bundler` parameter to the execution context (false by default).

We will have to address this properly in the future versions.